Выстроена структура проекта
