+++ .github/workflows/gh-pages.yml
@@ -10,13 +10,13 @@ jobs:
         with:
           node-version: 16
       - name: Install npm deps
-        working-directory: ./simple-chat
+        working-directory: ./react-chat
         run: npm i
       - name: Build app
-        working-directory: ./simple-chat
+        working-directory: ./react-chat
         run: npm run build
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: ./simple-chat/build
+          publish_dir: ./react-chat/build
