+++ react-chat/src/components/ChatListBody/ChatListBody.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+import restoreDialogues from '../../utils/restoreDialogues';
+
+export default function ChatListBody(props) {
+    return (
+        <div>
+            <a href="#">
+                <div className="new-message-btn">
+                    <i className="material-icons">edit</i>
+                </div>
+            </a>
+            <div className="message-list messages">
+                {restoreDialogues(props.props)}
+            </div>
+        </div>
+    )
+}
+
